fix(db): bind PostgreSQL parameters at inferred widths (#1156)
LIMIT/OFFSET and integer-literal comparisons bound i32 where PostgreSQL infers
bigint, and definition and evaluation-manifest advisory locks sent NUL-joined
text. Conformance harnesses now honor the test CA, race tests spawn before
joining, and scenarios that read across namespaces use scratch databases.
A source guard rejects the fixed bind shapes (#1155).

+++ src/db/postgres_decision.rs
@@ -83,12 +83,13 @@ impl PostgresDb {
         let action = filter.action.as_deref();
         let target_id = filter.target_id.as_deref();
         let after = filter.after;
-        let limit = if filter.limit > 0 {
+        // LIMIT and OFFSET parameters are inferred as bigint.
+        let limit = i64::from(if filter.limit > 0 {
             filter.limit
         } else {
             i32::MAX
-        };
-        let offset = filter.offset.max(0);
+        });
+        let offset = i64::from(filter.offset.max(0));
         self.connection()?
             .query(
                 "SELECT id,timestamp,actor,action,reason,evidence,target_id,outcome
